fix(activerecord): collectingQueriesForExplain return type; move composedOf into aggregations.ts (correct Rails location)
