+++ packages/activerecord/src/aggregations.ts
@@ -1,15 +1,15 @@
 import type { Base } from "./base.js";
 import { reload as persistenceReload } from "./persistence.js";
+import { AggregateReflection } from "./reflection.js";
 
 /**
- * Aggregation cache — memoizes composed-of value objects so repeated reads
- * return the same cached instance instead of allocating new ones each time.
+ * Aggregation cache and composed-of value-object support.
  *
  * Mirrors: ActiveRecord::Aggregations
  */
 
 // ---------------------------------------------------------------------------
-// Cache accessors used by composed-of.ts
+// Cache accessors
 // ---------------------------------------------------------------------------
 
 export function getAggregationCache(record: Base): Map<string, unknown> {
@@ -26,12 +26,101 @@ export function clearAggregationCache(record: Base): void {
 }
 
 // ---------------------------------------------------------------------------
-// Public instance methods
+// ClassMethods
+// ---------------------------------------------------------------------------
+
+interface ComposedOfOptions {
+  className: new (...args: any[]) => any;
+  mapping: [string, string][];
+  constructorFn?: (...args: any[]) => any;
+  converter?: (value: unknown) => unknown;
+}
+
+/**
+ * Configure a composed-of value object on a model.
+ *
+ * Mirrors: ActiveRecord::Aggregations::ClassMethods#composed_of
+ */
+export function composedOf(
+  modelClass: typeof Base,
+  name: string,
+  options: ComposedOfOptions,
+): void {
+  if (!Object.prototype.hasOwnProperty.call(modelClass, "_aggregateReflections")) {
+    const parent: Map<string, AggregateReflection> | undefined = (modelClass as any)
+      ._aggregateReflections;
+    (modelClass as any)._aggregateReflections = parent ? new Map(parent) : new Map();
+  }
+  (modelClass as any)._aggregateReflections.set(
+    name,
+    new AggregateReflection(
+      name,
+      null,
+      {
+        className: options.className.name,
+        mapping: options.mapping,
+        anonymousClass: options.className,
+      },
+      modelClass,
+    ),
+  );
+
+  Object.defineProperty(modelClass.prototype, name, {
+    get(this: Base): unknown {
+      const cache = getAggregationCache(this);
+      if (cache.has(name)) return cache.get(name);
+
+      const args = options.mapping.map(([modelAttr]) => this.readAttribute(modelAttr));
+      if (args.every((a) => a === null || a === undefined)) return null;
+
+      const obj = Object.freeze(new options.className(...args));
+      cache.set(name, obj);
+      return obj;
+    },
+    set(this: Base, value: unknown): void {
+      const cache = getAggregationCache(this);
+
+      if (value === null || value === undefined) {
+        for (const [modelAttr] of options.mapping) this.writeAttribute(modelAttr, null);
+        cache.delete(name);
+        return;
+      }
+
+      if (value instanceof options.className) {
+        for (const [modelAttr, valueAttr] of options.mapping)
+          this.writeAttribute(modelAttr, (value as any)[valueAttr]);
+        cache.set(
+          name,
+          Object.freeze(Object.assign(Object.create(Object.getPrototypeOf(value)), value)),
+        );
+        return;
+      }
+
+      if (options.converter) {
+        const converted = options.converter(value);
+        if (converted instanceof options.className) {
+          for (const [modelAttr, valueAttr] of options.mapping)
+            this.writeAttribute(modelAttr, (converted as any)[valueAttr]);
+          cache.set(
+            name,
+            Object.freeze(
+              Object.assign(Object.create(Object.getPrototypeOf(converted)), converted),
+            ),
+          );
+        }
+      }
+    },
+    configurable: true,
+  });
+}
+
+// ---------------------------------------------------------------------------
+// Instance methods
 // ---------------------------------------------------------------------------
 
 /**
  * Shallow-copy the aggregation cache into the duped record.
- * Cached value objects are frozen so sharing references across the dup is safe.
+ * Cached value objects are frozen so sharing references is safe.
  *
  * Mirrors: ActiveRecord::Aggregations#initialize_dup
  */
@@ -43,8 +132,7 @@ export function initializeDup(this: Base, _other: unknown): void {
 }
 
 /**
- * Clear the aggregation cache before reloading from the database so stale
- * value objects are not returned after the reload.
+ * Clear the aggregation cache before reloading from the database.
  *
  * Mirrors: ActiveRecord::Aggregations#reload
  */
@@ -57,7 +145,3 @@ export const InstanceMethods = {
   initializeDup,
   reload,
 };
-
-// Re-export composedOf so api:compare finds it under aggregations.ts
-// (Rails defines it in ActiveRecord::Aggregations::ClassMethods).
-export { composedOf } from "./composed-of.js";
